backend/requirements.txt
Django>=5.2
djangorestframework>=3.15
django-cors-headers>=4.3
pdfplumber>=0.11
python-docx>=1.1
google-genai>=1.0
python-dotenv>=1.0groq>=0.11